pyproject.toml
[project]
name = "claude-superpowers"
version = "0.1.0"
description = "Autonomous skill execution, messaging, scheduling, and orchestration for Claude Code"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "SmartMur", email = "contact@smartmur.com"}]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "cryptography>=42.0",
    "apscheduler>=3.10",
    "sqlalchemy>=2.0",
    "paramiko>=3.4",
    "watchdog>=4.0",
    "aiohttp>=3.9",
    "fastapi>=0.110",
    "PyJWT>=2.8",
    "uvicorn>=0.27",
    "redis>=5.0",
    "slack_sdk>=3.27",
    "playwright>=1.40",
    "httpx>=0.27",
    "mcp>=1.26.0",
    "anthropic>=0.39",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
]

[project.scripts]
claw = "superpowers.cli:main"
claw-mcp = "superpowers.mcp_server:main"

[tool.setuptools.packages.find]
include = ["superpowers*"]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["superpowers"]
